use setPerms from unix-compat
